+++ .pre-commit-config.yaml
@@ -10,20 +10,11 @@ repos:
     - id: check-symlinks
     - id: check-added-large-files
 
--   repo: https://github.com/myint/docformatter
-    rev: v1.4
+-   repo: https://gitlab.com/PyCQA/flake8
+    rev: 3.8.4
     hooks:
-    - id: docformatter
-
--   repo: https://github.com/myint/rstcheck
-    rev: '3f92957478422df87bd730abde66f089cc1ee19b'
-    hooks:
-    - id: rstcheck
-      args: [
-        "--report", "warning",
-        "--ignore-roles", "class",
-        "--ignore-directives", "autoclass,automodule",
-      ]
+    - id: flake8
+      args: ["--max-line-length=150"]
 
 -   repo: https://github.com/pre-commit/mirrors-autopep8
     rev: v1.5.4
@@ -48,8 +39,17 @@ repos:
         "--ignore", "E226,E24,W50,W690,E402"
       ]
 
--   repo: https://gitlab.com/PyCQA/flake8
-    rev: 3.8.4
+-   repo: https://github.com/myint/docformatter
+    rev: v1.4
     hooks:
-    - id: flake8
-      args: ["--max-line-length=150"]
+    - id: docformatter
+
+-   repo: https://github.com/myint/rstcheck
+    rev: '3f92957478422df87bd730abde66f089cc1ee19b'
+    hooks:
+    - id: rstcheck
+      args: [
+        "--report", "warning",
+        "--ignore-roles", "class",
+        "--ignore-directives", "autoclass,automodule",
+      ]
